Create Analyze Survey Data
In this project, I will analyze a real dataset about current Udacity students across a number of programs, so it isn't perfect. It is a little messy (some things are input incorrectly, others are missing). For this Project, I will use spreadsheets to analyze data and draw final conclusions based on the summary statistics.